refactor: ViewModelFactory 유틸 클래스를 di 패키지로 이동
